Reviewed-on: https://scm.openanalytics.eu/git/phaedra2-measurementservice/pulls/5

+++ server/src/main/java/eu/openanalytics/phaedra/measservice/repository/dao/MeasObjectStoreDAO.java
@@ -44,6 +44,7 @@ import com.amazonaws.services.s3.model.DeleteObjectsRequest;
 import com.amazonaws.services.s3.model.GetObjectRequest;
 import com.amazonaws.services.s3.model.ListObjectsV2Request;
 import com.amazonaws.services.s3.model.ListObjectsV2Result;
+import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.S3Object;
 import com.amazonaws.services.s3.model.S3ObjectInputStream;
 import com.amazonaws.services.s3.model.S3ObjectSummary;
@@ -66,6 +67,9 @@ public class MeasObjectStoreDAO {
 	
 	@Value("${meas-service.s3.upload-retry-delay:1000}")
 	private int uploadRetryDelayMs;
+	
+	@Value("${meas-service.s3.upload-temp-file-threshold:10000}")
+	private int uploadTempFileThreshold;
 	
 	@Value("${meas-service.s3.bucket-name}")
 	private String bucketName;
@@ -146,9 +150,13 @@ public class MeasObjectStoreDAO {
 	public void putMeasObjectRaw(long measId, String key, byte[] value) throws IOException {
 		String s3key = makeS3Key(measId, key);
 		
-		File tempFile = File.createTempFile("meas-data", null);
-		try (OutputStream out = new FileOutputStream(tempFile)) {
-			StreamUtils.copy(value, out);
+		boolean useTempFile = value.length > uploadTempFileThreshold;
+		File tempFile = null;
+		if (useTempFile) {
+			tempFile = File.createTempFile("meas-data", null);
+			try (OutputStream out = new FileOutputStream(tempFile)) {
+				StreamUtils.copy(value, out);
+			}
 		}
 		
 		// By using a file instead of a stream, TransferManager can optimize the upload: multipart parallel uploads with auto-retrying.
@@ -158,7 +166,14 @@ public class MeasObjectStoreDAO {
 			int currentTry = 1;
 			while (currentTry <= uploadMaxTries) {
 				try {
-					Upload upload = transferMgr.upload(bucketName, s3key, tempFile);
+					Upload upload = null;
+					if (useTempFile) {
+						upload = transferMgr.upload(bucketName, s3key, tempFile);
+					} else {
+						ObjectMetadata mtdt = new ObjectMetadata();
+						mtdt.setContentLength(value.length);
+						upload = transferMgr.upload(bucketName, s3key, new ByteArrayInputStream(value), mtdt);
+					}
 					upload.waitForCompletion();
 					return;
 				} catch (Exception e) {
@@ -170,7 +185,7 @@ public class MeasObjectStoreDAO {
 				}
 			}
 		} finally {
-			tempFile.delete();
+			if (useTempFile) tempFile.delete();
 		}
 		throw new IOException(String.format("Failed to upload data to S3 for meas %d and key %s", measId, key), caughtException);		
 	}
